requirements.txt
weaviate-client
sentence-transformers
python-docx
PyMuPDF
deep-translator
langdetect
